Clean up utility exports
